requirements.txt
tensorflow==2.11.0
biopython==1.80
matplotlib==3.6.2
